+++ ShuInkWeb/ShuInkWeb/Controllers/BlogController.cs
@@ -3,9 +3,8 @@ using Microsoft.AspNetCore.Mvc;
 using ShuInkWeb.Core.Contracts;
 using ShuInkWeb.Core.Models.HappeningModels;
 using ShuInkWeb.Controllers.Common;
-using Ganss.Xss;
-using CloudinaryDotNet.Actions;
-using CloudinaryDotNet;
+using static ShuInkWeb.Constants.ActionsConstants;
+using static ShuInkWeb.Constants.AreaConstants;
 
 namespace ShuInkWeb.Controllers
 {
@@ -23,27 +22,55 @@ namespace ShuInkWeb.Controllers
         {
             var models = await happeningService.GetHappeningsAsync();
 
+            if (!(models.Any()))
+            {
+                return RedirectToAction(IndexConst, HomeConst);
+            }
+
             return View(models);
         }
 
         [HttpGet]
+        [Authorize(Roles = ArtistRoleName)]
         public IActionResult Add([FromForm] IFormFile file)
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             var model = new HappeningViewModel();
 
             return View(model);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Add([FromForm] IFormFile file,HappeningViewModel model)
+        [Authorize(Roles = ArtistRoleName)]
+        public async Task<IActionResult> Add([FromForm] IFormFile file, HappeningViewModel model)
         {
-            await happeningService.AddHappeningAsync(model,file);
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
+            if (!ModelState.IsValid)
+            {
+                return View(model);
+            }
+
+            await happeningService.AddHappeningAsync(model, file);
 
             return RedirectToAction(nameof(All));
         }
 
+        [Authorize(Roles = ArtistRoleName)]
         public async Task<IActionResult> Details(Guid id)
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             if (await happeningService.HappeningExist(id))
             {
                 RedirectToAction(nameof(All));
@@ -55,14 +82,25 @@ namespace ShuInkWeb.Controllers
         }
 
         [HttpGet]
+        [Authorize(Roles = ArtistRoleName)]
         public async Task<IActionResult> Edit(Guid id)
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             if ((await happeningService.HappeningExist(id) == false))
             {
                 return RedirectToAction(nameof(All));
             }
 
             var post = await happeningService.GetSingleHappeningAsync(id);
+
+            if (post == null)
+            {
+                return RedirectToAction(nameof(All));
+            }
 
             var model = new HappeningViewModel()
             {
@@ -76,19 +114,27 @@ namespace ShuInkWeb.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = ArtistRoleName)]
         public async Task<IActionResult> Edit(Guid id, HappeningViewModel model)
         {
-            if (id != model.Id)
-            {
-                return RedirectToPage("/Account/AccessDenied", new { area = "Identity" });
-            }
-            if ((await happeningService.HappeningExist(id) == false))
-            {
-                return RedirectToAction(nameof(All));
-            }
             if (!ModelState.IsValid)
             {
                 return View(model);
+            }
+
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
+            if (id != model.Id)
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
+            if ((await happeningService.HappeningExist(id) == false))
+            {
+                return RedirectToAction(nameof(All));
             }
 
             await happeningService.Edit(id, model);
@@ -96,8 +142,19 @@ namespace ShuInkWeb.Controllers
             return RedirectToAction(nameof(All));
         }
 
+        [Authorize(Roles = ArtistRoleName)]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
+            if ((await happeningService.HappeningExist(id) == false))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             await happeningService.Delete(id);
 
             return RedirectToAction(nameof(All));
